+++ apps/compartmentConverter.cpp
@@ -47,6 +47,8 @@ using boost::lexical_cast;
         ::exit( EXIT_FAILURE );                         \
     }
 
+namespace
+{
 template< class T > void requireEqualCollections( const T& a, const T& b )
 {
     typename T::const_iterator i = a.begin();
@@ -59,6 +61,30 @@ template< class T > void requireEqualCollections( const T& a, const T& b )
     }
     REQUIRE_EQUAL( i, a.end( ));
     REQUIRE_EQUAL( j, b.end( ));
+}
+
+/** @return true if the cell occupies a continuous region in the report frame */
+bool _isCompact( const brion::CompartmentReport& report,
+                    const size_t gidIndex )
+{
+    const auto& offsets = report.getOffsets()[gidIndex];
+    const auto& counts = report.getCompartmentCounts()[gidIndex];
+
+    // sections are not guaranteed to be enumerated in order - sort them:
+    std::map< uint64_t, uint16_t > mapping;
+    for( size_t i = 0; i < offsets.size(); ++i )
+        mapping.emplace(offsets[i], counts[i]);
+
+    auto i = mapping.begin();
+    uint64_t next = i->first + i->second;
+    for( ++i; i != mapping.end(); ++i )
+    {
+        if( i->first != next )
+            return false;
+        next = i->first + i->second;
+    }
+    return true;
+}
 }
 
 /**
@@ -90,7 +116,7 @@ int main( const int argc, char** argv )
 #else
         ( "input,i", po::value< std::string >()->required(), "Input report URI")
 #endif
-        ( "output,o", po::value< std::string >()->default_value( "null://" ),
+        ( "output,o", po::value< std::string >()->default_value( "dummy://" ),
           uriHelp.c_str( ))
         ( "erase,e", po::value< std::string >(),
           "Erase the given report (map-based reports only)" )
@@ -193,14 +219,13 @@ int main( const int argc, char** argv )
     clock.reset();
     brion::CompartmentReport to( outURI, brion::MODE_OVERWRITE );
     to.writeHeader( start, end, step, in.getDataUnit(), in.getTimeUnit( ));
-
     {
         size_t index = 0;
         for( const uint32_t gid : gids )
-            to.writeCompartments( gid, counts[ index++ ] );
+            if( !to.writeCompartments( gid, counts[ index++ ] ))
+                return EXIT_FAILURE;
     }
 
-    to.flush(); // write header before parallel section
     float writeTime = clock.getTimef();
 
     const size_t nFrames = (end - start) / step;
@@ -218,24 +243,36 @@ int main( const int argc, char** argv )
             ::exit( EXIT_FAILURE );
         }
 
-        const brion::floats& voltages = *data.get();
-        const brion::SectionOffsets& offsets = in.getOffsets();
+        const brion::floats& values = *data.get();
+        const auto& offsets = in.getOffsets();
 
         size_t index = 0;
         clock.reset();
         for( const uint32_t gid : gids )
         {
-            brion::floats cellVoltages;
-            cellVoltages.reserve( in.getNumCompartments( index ));
+            if( _isCompact( in, index ))
+            {
+                const float* cellValues = &values[ offsets[index][0] ];
+                const size_t size = std::accumulate( counts[index].begin(),
+                                                     counts[index].end(), 0 );
+                if( !to.writeFrame( gid, cellValues, size, t ))
+                    return EXIT_FAILURE;
+                ++index;
+                continue;
+            }
+
+            brion::floats cellvalues;
+            cellvalues.reserve( in.getNumCompartments( index ));
 
             for( size_t j = 0; j < offsets[index].size(); ++j )
             {
                 const auto offset = offsets[index][j];
                 for( size_t k = 0; k < counts[index][j]; ++k )
-                    cellVoltages.emplace_back( voltages[ offset + k ] );
+                    cellvalues.emplace_back( values[ offset + k ] );
             }
 
-            to.writeFrame( gid, cellVoltages, t );
+            if( !to.writeFrame( gid, cellvalues, t ))
+                return EXIT_FAILURE;
             ++index;
         }
         writeTime += clock.getTimef();
@@ -248,7 +285,8 @@ int main( const int argc, char** argv )
 
     std::cout << "Converted " << inURI << " to " << outURI
               << " (in " << size_t( loadTime ) << " out " << size_t( writeTime )
-              << " ms)" << std::endl;
+              << " ms, " << gids.size() << " cells X " <<  nFrames << " frames)"
+              << std::endl;
 
     if( vm.count( "compare" ))
     {